feat: take out is frequently used button component
