+++ src/components/FrequentlyUsedFunctions/FrequentlyUsedButtons/FrequentlyUsedButton/frequentlyUsedButton.settings.ts
@@ -1,0 +1,4 @@
+import { colors } from '../../../../constants/colors';
+
+export const iconColor = colors.white;
+export const iconSize = 32;
