+++ src/components/FrequentlyUsedFunctions/FrequentlyUsedButtons/FrequentlyUsedButton/frequentlyUsedButton.styles.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from 'react-native';
+
+import { colors } from '../../../../constants/colors';
+
+const styles = StyleSheet.create({
+  frequentlyButton: {
+    borderRadius: 50,
+    padding: 10,
+    backgroundColor: colors.darkGrey,
+    marginBottom: 5,
+  },
+  center: {
+    alignItems: 'center',
+  },
+});
+
+export default styles;
