+++ src/components/FrequentlyUsedFunctions/FrequentlyUsedButtons/frequentlyUsedButtons.settings.ts
@@ -1,8 +1,3 @@
-import { colors } from '../../../constants/colors';
-
-export const iconColor = colors.white;
-export const iconSize = 32;
-
 export const transferButtonTitle = 'Transfer';
 export const transferButtonIcon = 'arrowup';
 export const transferButtonOnPress = () => console.log('Transfer');
@@ -14,3 +9,21 @@ export const depositButtonOnPress = () => console.log('Deposit');
 export const detailsButtonTitle = 'Details';
 export const detailsButtonIcon = 'bars';
 export const detailsButtonOnPress = () => console.log('Details');
+
+export const buttons = [
+  {
+    title: transferButtonTitle,
+    iconName: transferButtonIcon,
+    onPress: transferButtonOnPress,
+  },
+  {
+    title: depositButtonTitle,
+    iconName: depositButtonIcon,
+    onPress: depositButtonOnPress,
+  },
+  {
+    title: detailsButtonTitle,
+    iconName: detailsButtonIcon,
+    onPress: detailsButtonOnPress,
+  },
+];
